Lint and truffle config cleanup
